pyproject.toml
[project]
name = "discord-hack"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiohttp>=3.8.0",
    "discord-py>=2.6.3",
    "openai>=1.109.1",
    "pydantic-ai>=1.0.10",
    "pydantic-evals[logfire]>=1.0.10",
    "python-dotenv>=1.1.1",
]

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[dependency-groups]
dev = [
    "pyright>=1.1.405",
    "pytest>=8.4.2",
    "pytest-asyncio>=1.2.0",
    "ruff>=0.13.2",
    "trio>=0.31.0",
]

[tool.ruff.lint]
extend-select = [
  "UP",  # pyupgrade
]
